use vcpkg directly from within CMake
this makes the experience much better as the install doesn't have to be
done manually anymore. instead, the dependencies are managed via
`vcpkg.json` and the integrated toolchain takes care of everything.

to integrate the toolchain via the env. variable CMake >= 3.21 is
needed.

thus the VS-specific config file is also no longer needed.