simplify GHA required build settings
by adding a single `build-results` job which depends on all other jobs
we can simplify the setting of required builds in the repository.
currently, all builds - including all variations of the build matrix! -
need to be manually specified.

once this has been merged the settings can be changed to require only
this one job (which will fail if any of the other jobs failed).

this way it's also easier to add/remove jobs or change the build matrix
as it no longer requires changing the settings on the repository.

this is inspired by [this discussion on GH][discussion].

[discussion]: https://github.com/orgs/community/discussions/26822