add `CMakePresets.json`
this adds the possibility to use presets when running CMake.
see [the CMake documentation][cmake-presets] for more information.

this is also a precondition to using the [run-cmake] GitHub action in a
subsequent step.

in order to use the presets you must have an environment variable named
`VCPKG_ROOT` pointing to the root directory of your vcpkg installation.

[cmake-presets]: https://cmake.org/cmake/help/latest/manual/cmake-presets.7.html
[run-cmake]: https://github.com/marketplace/actions/run-cmake